changing detection return value

+++ detection/detection.h
@@ -1,5 +1,5 @@
 #ifndef DETECTIONS_H
 #define DETECTION_H
 
-int** detection(int* mat, int w, int h);
+double** detection(int* mat, int w, int h);
 #endif
